+++ web/src/components/operate-dropdown/index.less
@@ -1,0 +1,3 @@
+.delete {
+  height: 24px;
+}
